tests/conftest.py: tox: Remove root dir from sys.path
When using tox, it can accidentally pick up a {pymssql,_mssql}.so file
in the root directory and then get ImportError because of
incompatibility in Python versions. By removing the root directory from
the sys.path, it forces tox to import the library from correct place in
the tox virtualenv.